move flatwing to wingloads
