+++ wingstructure/data/wing.py
@@ -259,57 +259,3 @@ class Wing(_Wing):
         plt.axis('off')
         plt.xlim(-1, max(y_positions)+1)
 
-            
-class FlatWing(Wing):
-    """A class representing the flattend version of a wing
-
-    Can be useful in calculation of aerodynamic loads with methods not suporting
-    dihedral. Flat means parallel to the x-y-plane.
-    
-    Parameters
-    ----------
-    wing : Wing
-        the wing to be flattend
-    """
-    
-    def __init__(self, wing):
-        """create flat wing instance
-        
-        Returns
-        -------
-        Wing
-            a flat wing instance
-        """
-
-        super().__init__((wing.x, wing.y, wing.z))
-
-        self.basewing = wing
-
-        lastsec = None
-        lastpos = 0.0
-
-        for section in wing.sections:
-
-            curpos = section.pos
-
-            if lastsec is not None:
-                # calculate section distance in y-z-plane
-                pos1 = np.array(section.pos)
-                pos2 = np.array(lastsec.pos)
-
-                # ignore x -> 2D vector
-                distvec = (pos1-pos2)[(1,2),] 
-                
-                dist = np.linalg.norm(distvec)
-
-                newpos = curpos._replace(y=lastpos+dist, z=0.0)
-                newsec = section._replace(pos=newpos)
-                
-            else:
-                # set z to zero
-                newpos = curpos._replace(z=0.0)
-                newsec = section._replace(pos=newpos)
-            
-            self.sections.append(newsec)
-            lastsec = section
-            lastpos = newsec.pos[1]
